use test coverage to fix bug with existing sessions
the cookie reader was not compiling a regexp, and was just returning the whole
cookie header. this SOMETIMES worked, if no other cookies existed or contained
periods, because then the split('.')[1] might still happen to grab the payload.

but this is more reliable. hooray for tests!

other changes: make the session available publicly. this is necessary for tests
but should be used with caution elsewhere, since the token is yet to be trusted.