consolidate UMD builds
after https://github.com/keratin/authn-js/pull/6, the persistence stores are fairly small. so
this removes the quasi-"tree shaking" that the various UMD bundles offered. vanilla JS stacks
are still supported, but that level of micro-optimization for vanilla JS is not.

the total file size change for keratin-authn.min.js is ~1.2kb, up from 4.9kb to 6.1kb.