refactor logout() to use jhr()
the iframe was implemented at a time where i thought that logouts would have a redirection
to the host app for any cleanup tasks there. but once i realized that the promise chain was
sufficient, i never even added the ability to specify a redirect param and the iframe was
underutilized.

fixes #10