recalculate refresh interval on resume from sleep
timers can get weird when devices sleep. i've observed timers that run
much later than expected after waking up the device. this creates a
period of time where the device thinks it has a session but it does not.

in the interest of minimizing dependency on the device clock, we set a
refreshAt target so we can recalculate the new interval when resuming.